Revert "Introduce pagination for movies page"
