+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.61",
     "@types/body-scroll-lock": "^3.1.0",
     "axios": "^1.2.2",
     "axios-case-converter": "^0.11.1",
