+++ pages/movies/index.tsx
@@ -1,0 +1,27 @@
+import { NextPage } from "next";
+import { MovieList } from "../../src/domain/movies/model";
+import { getMyFavoriteMovies } from "../../src/data/movies/index";
+import { getGenres } from "../../src/data/movies/index";
+import MoviePageComponent from "src/components/pages/movies";
+import { Genre } from "src/domain/genres/model";
+
+type PageProps = {
+  myFavoriteMovieList: MovieList;
+  genres: Array<Genre>;
+};
+
+const MoviePage: NextPage<PageProps> = ({ myFavoriteMovieList, genres }) => {
+  return <MoviePageComponent movieList={myFavoriteMovieList} genres={genres} />;
+};
+
+export default MoviePage;
+
+export const getStaticProps = async () => {
+  const genres = await getGenres();
+
+  const myFavoriteMovieList = await getMyFavoriteMovies(genres);
+
+  return {
+    props: { myFavoriteMovieList, genres },
+  };
+};
