--- src/data/movies/constants.ts
@@ -1,1 +1,0 @@
-export const MY_FAVORITE_MOVIES_COUNT_PER_PAGE = 20;
